Intersect unit test fix
EuclideanIntersect is not aware of camera calibration
matrix yet and always assumes it to be an identity.

So using non-identity matrix to construct sample case
leads to wrong projection results.

For now made it so test case uses identity matrix for
calibration.

Also fixed variable shadowing which lead to wrong
markers positions (were either zero or undefined).