Pyramid tracker unit test fix
Issue was caused by trackers modifying guessed
point location even in case of failure. So made
id so both level 0 and level 3 of pyramid are
starting from the same initial guessed location.

Modifying locations in case tracker returns false
is not actually a bug - someone could still want
to use that location. False in this case  means
more like "returned location is not so much
accurate".