Revert "Change libmv's bilinear sampling to assume the same"
Revert changes to bilinear sampler which were originally
aimed to match blender's pixel center (where integer coord
is a left-bottom corner, x.5 coords are centers.

The reason of revert is changing this assumption in only
sampler didn't work well and lead to wrong results of
BlurredImageAndDerivativesChannels for example.

Discovered when was doing unit-tests for brute region tracker.

This reverts commit daa354c0735b954b0cd7725626e9a3d67416d46b.