Unit tests for feature detector
Currently covers only simplest cases with synthetic images.
Also at this point mainly Harris detector is being testes,
other detectors behaves a bit unexpected on synthetic images
and this is to be investigated further.

Tests will be extended further later.

Additional change:

- Added constructor to Feature structure
- Added operator << for feature for easier debug dumps.

TODO: Some tests are not giving the result which i was expected
to. This is to be investigated further by finding the reference
detector implementation. For until then keeping that tests
commented out.

Reviewers: keir

Reviewed By: keir

Differential Revision: https://developer.blender.org/D316