Fix homography test failure
It was caused by assuming that reconstructed homography matrix
should look exactly the same as the matrix used to generate a
test case.

It's not actually valid assumption because different-looking
matrices could correspond to the same exact transform.

In this change we make it so actual "re-projected" vectors
are being checked, not the values in matrix. This makes it
more predictable verification.

Reviewers: keir

Reviewed By: keir

Differential Revision: https://developer.blender.org/D488