Initial commit of unfinished AutoTrack API
This starts the creating the new AutoTrack API. The new API will
make it possible for libmv to do full autotracking, including
predictive tracking and also support multiple motion models (3D
planes etc).

The first goal (not in this patch) is to convert Blender to use
the new API without adding any new functionality.

Note: This does not add any of the API to the build system!
It likely does not compile.

Reviewers: sergey

Reviewed By: sergey

Differential Revision: https://developer.blender.org/D499