Update GFlags to latest release 2.1.1
Main purpose of this (andsome of upcoming) update is to check if the
upstream sources are useable without any modifications for us. And if
not, then we'll need to consider moving some changes into upstream.

This commit contains an one-to-one copy of the upstream GFlags library
and also changes namespace usage since it's changed in the upstream.